outputstream file modularization
