+++ src/stx/io/FileOutputStream.h
@@ -1,0 +1,91 @@
+/**
+ * This file is part of the "libstx" project
+ *   Copyright (c) 2014 Paul Asmuth, Google Inc.
+ *
+ * libstx is free software: you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License v3.0. You should have received a
+ * copy of the GNU General Public License along with this program. If not, see
+ * <http://www.gnu.org/licenses/>.
+ */
+#ifndef _libstx_io_FileOutputStream_h
+#define _libstx_io_FileOutputStream_h
+#include <fcntl.h>
+#include <stx/io/outputstream.h>
+#include <stx/io/file.h>
+
+namespace stx {
+
+class FileOutputStream : public OutputStream {
+public:
+  /**
+   * Create a new FileOutputStream instance by opening the provided file for
+   * writing. The fille will automatically be close()ed when the output stream
+   * is destroyed.
+   *
+   * @param file_path the path to the file to open
+   * @param flags flags to pass to the open() syscall
+   */
+  static std::unique_ptr<FileOutputStream> openFile(
+      const std::string& file_path,
+      int flags = O_CREAT | O_TRUNC,
+      int permissions = 0666);
+
+  /**
+   * Create a new FileOutputStream instance from the provided filedescriptor. If
+   * close on_destroy is true, the fd will be close()ed when the input stream
+   * is destroyed.
+   *
+   * @param fd a valid fd
+   * @param close_on_destroy close the fd on destroy?
+   */
+  static std::unique_ptr<FileOutputStream> fromFileDescriptor(
+      int fd,
+      bool close_on_destroy = false);
+
+  /**
+   * Create a new FileOutputStream instance from the provided File.
+   *
+   * @param file the opened file
+   */
+  static std::unique_ptr<FileOutputStream> fromFile(File&& file);
+
+  /**
+   * Create a new FileOuputStream instance from the provided filedescriptor. If 
+   * close on_destroy is true, the fd will be close()ed when the output stream
+   * is destroyed.
+   *
+   * @param fd a valid fd
+   * @param close_on_destroy close the fd on destroy?
+   */
+  explicit FileOutputStream(int fd, bool close_on_destroy = false);
+
+  /**
+   * Create a new FileOutputStream instance from the provided File.
+   *
+   * @param file the opened file
+   */
+  explicit FileOutputStream(File&& file);
+
+  /**
+   * Close the fd if close_on_destroy is true
+   */
+  ~FileOutputStream();
+
+  /**
+   * Write the next n bytes to the file. This may raise an exception.
+   * Returns the number of bytes that have been written.
+   *
+   * @param data a pointer to the data to be written
+   * @param size then number of bytes to be written
+   */
+  size_t write(const char* data, size_t size) override;
+
+  size_t printf(const char* format, ...) override;
+
+protected:
+  int fd_;
+  bool close_on_destroy_;
+};
+
+} // namespace stx
+#endif
